requirements.txt
numpy
tensorflow
Pillow
tk
keras